
\name{interflex-internal}
\alias{interflex.plot.pool}
\alias{interflex.linear}
\alias{interflex.binning}
\alias{interflex.kernel}
\alias{interflex.gam}
\alias{interflex.raw}
\alias{vcovCluster} 
\alias{fastplm}
\alias{iv_fastplm}
\title{Internal Interflex Functions}
\description{
 Internal Interflex functions
}
\details{
These are not to be called by the user, or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ts}
