
\name{interflex-internal}
\alias{coefs}
\alias{crossvalidate}
\alias{fastplm}
\alias{fastplm_predict}
\alias{interflex_fastplm}
\alias{interflex_fastplm_predict}
\alias{vcovCluster} 
\title{Internal Interflex Functions}
\description{
 Internal Interflex functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ts}
